+++ packages/web3/test/e2e/e2e_utils.ts
@@ -51,8 +51,9 @@ export const getE2ETestAccountAddress = (): string => {
 };
 
 export const getE2ETestContractAddress = () =>
+	// eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
 	secrets[getSystemTestBackend().toUpperCase() as 'SEPOLIA' | 'MAINNET']
-		.DEPLOYED_TEST_CONTRACT_ADDRESS;
+		.DEPLOYED_TEST_CONTRACT_ADDRESS as string;
 
 export const getAllowedSendTransaction = (): boolean => {
 	if (process.env.ALLOWED_SEND_TRANSACTION !== undefined) {