+++ packages/web3/CHANGELOG.md
@@ -112,8 +112,19 @@ Documentation:
 [Web3 API](https://docs.web3js.org/api)
 [Migration Guide from 1.x](https://docs.web3js.org/guides/web3_upgrade_guide/x/)
 
-## [Unreleased]
+## [4.0.2]
 
 ### Added
 
 -   Exported `Web3Context`, `Web3PluginBase`, `Web3EthPluginBase` from `'web3-core'`, and `Web3Validator` from `'web3-validator'` (#6165)
+
+### Fixed
+
+-   Fixed bug #6185, now web3.js compiles on typescript v5 (#6195)
+-   Fixed #6162 @types/ws issue (#6205)
+
+## [Unreleased]
+
+### Fixed
+
+-   Fixed bug #6236 by adding personal type in web3.eth (#6245)